+++ src/fieldlayoutelements/MetafieldsField.php
@@ -74,30 +74,12 @@ class MetafieldsField extends BaseNativeField
             ];
         }
 
-        // @TODO remove this when Craft 4 support is dropped
-        $name = method_exists($this, 'baseInputName') ? $this->baseInputName() : $this->attribute();
-
-        $tableConfig = [
+        return Cp::editableTableHtml([
             'id' => $this->id(),
-            'name' => $name,
+            'name' => $this->baseInputName(),
             'cols' => $cols,
             'rows' => $tableData,
             'static' => true,
-        ];
-
-        return $this->_editableTableHtml($tableConfig);
-    }
-
-    /**
-     * @TODO remove this when Craft 4 support is dropped
-     */
-    private function _editableTableHtml(array $tableConfig): string
-    {
-        // @phpstan-ignore booleanNot.alwaysFalse (Craft 4 compatibility: method does not exist in Craft 4)
-        if (!is_callable([Cp::class, 'editableTableHtml'])) {
-            return Cp::renderTemplate('_includes/forms/editableTable.twig', $tableConfig);
-        }
-
-        return Cp::editableTableHtml($tableConfig);
+        ]);
     }
 }